requirements.txt
# 核心依赖
psycopg2-binary
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.12
tiktoken==0.5.2
redis==5.0.1

# 文档处理
pypdf==3.17.4
python-docx

# Web服务
flask

# 安全相关
cryptography
passlib
PyJWT

# 数据科学和向量化
numpy==1.24.4
pandas==2.1.4

# 网络请求
requests==2.31.0
urllib3==2.1.0

# 其他工具
python-dotenv==1.0.0
tqdm==4.66.1